AMDA: some cleanups plus introduced DatasetIndex (same idea than TimetableIndex)
