+++ speasy/amda/timetable.py
@@ -1,17 +1,10 @@
 """AMDA Timetable definition. Serves a cosmetic need for now, functionality to be added in the future.
 """
-import numpy as np
 
 from speasy.common.variable import SpeasyVariable
 
 
-class TimetableIndex:
-    def __init__(self, uid, name):
-        self.uid = uid
-        self.name = name
 
-    def __repr__(self):
-        return f'<TimetableIndex: {self.name}, id: {self.uid}>'
 
 
 class TimeTable(SpeasyVariable):
